jreader: add StringAsBytes() (#19)
When reading a new version of a large but shallow JSON file from disk
(e.g. an array of feature flag definitions), we end up creating a lot of
strings for object properties (e.g. FF name), just to end up eventually
comparing those strings to the same properties on an already-read-in
older version of a flag, and discarding the newly created ones. Exposing
`StringAsBytes()` as an alternative API to `String()` enables us to
compare new versions of flags (at Stripe we use a pre-existing, internal
fork of a feature flag library called
[`goforit`](https://github.com/stripe/goforit) to old ones without
creating garbage, as due to quirks of our distribution mechanism the
majority of flags rarely change.

Using `go-jsonstream` with this change (together with
https://github.com/launchdarkly/go-jsonstream/pull/18 , but this PR is
the most impactful) results in a 60% reduction in runtime in my usecase
(updating our feature flags from disk) compared to using
`encoding/json`, an 80% reduction in MB allocated, and a 99%(!)
reduction in objects allocated.